+++ doc/versionHistory.rst
@@ -4,6 +4,16 @@
 Version History
 ##################
 
+.. _lsst.ts.donut.viz-1.1.0:
+
+-------------
+1.1.0
+-------------
+
+* Add tests for full donut_viz pipeline.
+* Add changelog github action.
+* Add Jenkinsfile.
+
 .. _lsst.ts.donut.viz-1.0.0:
 
 -------------